backend/requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
certifi==2019.9.11
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
Cython==0.29.23
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.4
django-cors-headers==3.7.0
django-recaptcha==2.0.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djantic==0.3.3
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
Mako==1.1.4.dev0
Markdown @ file:///buildstream/freedesktop-sdk/components/python3-markdown.bst
meson @ file:///buildstream/freedesktop-sdk/components/meson.bst
oauthlib==3.1.1
Pillow==8.2.0
pipenv==2018.11.26
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0.dev20210520
PyJWT==2.1.0
pyOpenSSL==20.0.1
python-dotenv==0.17.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.5
virtualenv==16.7.7
virtualenv-clone==0.5.3
zope.interface==5.4.0
